Fixed bug in GetNextOccurrence Function
Changed GetServicesFunction to return all services with a date inside the number requested. For example: If your request 100 and the 100th service's date is 3-26-2012, but there are another 10 services on that date, the function will return 110 services instead of 100.

Began to further optimize GetServicesFunction - down to ~1.9 seconds for 1000 services when requesting both previous and future and ~1.3 seconds for 1000 services if you request previous or future